# cssbattle-server/models/InterfaceEntity/user.go
package InterfaceEntity

//汇总实体类
type UserInfo struct {
	UserName     string  `json:"userName"`
	UserIcon     string  `json:"userIcon"`
	Password     string  `json:"password"`
	CreateTime   string  `json:"createTime"`
	ModifyTime   string  `json:"modifyTime"`
	Score        float32 `json:"score"`
	Target       int     `json:"target"`
	Cssbattle_1  float32 `json:"cssbattle_1"`
	Cssbattle_2  float32 `json:"cssbattle_2"`
	Cssbattle_3  float32 `json:"cssbattle_3"`
	Cssbattle_4  float32 `json:"cssbattle_4"`
	Cssbattle_5  float32 `json:"cssbattle_5"`
	Cssbattle_6  float32 `json:"cssbattle_6"`
	Cssbattle_7  float32 `json:"cssbattle_7"`
	Cssbattle_8  float32 `json:"cssbattle_8"`
	Cssbattle_9  float32 `json:"cssbattle_9"`
	Cssbattle_10 float32 `json:"cssbattle_10"`
	Cssbattle_11 float32 `json:"cssbattle_11"`
	Cssbattle_12 float32 `json:"cssbattle_12"`
	Cssbattle_13 float32 `json:"cssbattle_13"`
	Cssbattle_14 float32 `json:"cssbattle_14"`
	Cssbattle_15 float32 `json:"cssbattle_15"`
	Cssbattle_16 float32 `json:"cssbattle_16"`
	Cssbattle_17 float32 `json:"cssbattle_17"`
	Cssbattle_18 float32 `json:"cssbattle_18"`
	Cssbattle_19 float32 `json:"cssbattle_19"`
	Cssbattle_20 float32 `json:"cssbattle_20"`
	Cssbattle_21 float32 `json:"cssbattle_21"`
	Cssbattle_22 float32 `json:"cssbattle_22"`
	Cssbattle_23 float32 `json:"cssbattle_23"`
	Cssbattle_24 float32 `json:"cssbattle_24"`
	Cssbattle_25 float32 `json:"cssbattle_25"`
	Cssbattle_26 float32 `json:"cssbattle_26"`
	Cssbattle_27 float32 `json:"cssbattle_27"`
	Cssbattle_28 float32 `json:"cssbattle_28"`
	Cssbattle_29 float32 `json:"cssbattle_29"`
	Cssbattle_30 float32 `json:"cssbattle_30"`
	Cssbattle_31 float32 `json:"cssbattle_31"`
	Cssbattle_32 float32 `json:"cssbattle_32"`
	Cssbattle_33 float32 `json:"cssbattle_33"`
	Cssbattle_34 float32 `json:"cssbattle_34"`
	Cssbattle_35 float32 `json:"cssbattle_35"`
	Cssbattle_36 float32 `json:"cssbattle_36"`
	Cssbattle_37 float32 `json:"cssbattle_37"`
	Cssbattle_38 float32 `json:"cssbattle_38"`
	Cssbattle_39 float32 `json:"cssbattle_39"`
	Cssbattle_40 float32 `json:"cssbattle_40"`
	Cssbattle_41 float32 `json:"cssbattle_41"`
	Cssbattle_42 float32 `json:"cssbattle_42"`
	Cssbattle_43 float32 `json:"cssbattle_43"`
	Cssbattle_44 float32 `json:"cssbattle_44"`
	Cssbattle_45 float32 `json:"cssbattle_45"`
	Cssbattle_46 float32 `json:"cssbattle_46"`
	Cssbattle_47 float32 `json:"cssbattle_47"`
	Cssbattle_48 float32 `json:"cssbattle_48"`
	Cssbattle_49 float32 `json:"cssbattle_49"`
	Cssbattle_50 float32 `json:"cssbattle_50"`
	Cssbattle_51 float32 `json:"cssbattle_51"`
	Cssbattle_52 float32 `json:"cssbattle_52"`
	Cssbattle_53 float32 `json:"cssbattle_53"`
	Cssbattle_54 float32 `json:"cssbattle_54"`
	Cssbattle_55 float32 `json:"cssbattle_55"`
	Cssbattle_56 float32 `json:"cssbattle_56"`
	Cssbattle_57 float32 `json:"cssbattle_57"`
	Cssbattle_58 float32 `json:"cssbattle_58"`
	Cssbattle_59 float32 `json:"cssbattle_59"`
	Cssbattle_60 float32 `json:"cssbattle_60"`
	Cssbattle_61 float32 `json:"cssbattle_61"`
	Cssbattle_62 float32 `json:"cssbattle_62"`
	Cssbattle_63 float32 `json:"cssbattle_63"`
	Cssbattle_64 float32 `json:"cssbattle_64"`
	Cssbattle_65 float32 `json:"cssbattle_65"`
	Cssbattle_66 float32 `json:"cssbattle_66"`
	Cssbattle_67 float32 `json:"cssbattle_67"`
	Cssbattle_68 float32 `json:"cssbattle_68"`
	Cssbattle_69 float32 `json:"cssbattle_69"`
	Cssbattle_70 float32 `json:"cssbattle_70"`
	Cssbattle_71 float32 `json:"cssbattle_71"`
	Cssbattle_72 float32 `json:"cssbattle_72"`
	Cssbattle_73 float32 `json:"cssbattle_73"`
	Cssbattle_74 float32 `json:"cssbattle_74"`
	Cssbattle_75 float32 `json:"cssbattle_75"`
	Cssbattle_76 float32 `json:"cssbattle_76"`
	Cssbattle_77 float32 `json:"cssbattle_77"`
	Cssbattle_78 float32 `json:"cssbattle_78"`
	Cssbattle_79 float32 `json:"cssbattle_79"`
	Cssbattle_80 float32 `json:"cssbattle_80"`
}